Major code clean up
More still to do!

5.9 FPS vs 7.9 FPS originally

+++ pmma/core/hpp_src/Rendering/Shapes2D/ArcShape.hpp
@@ -26,7 +26,7 @@ class EXPORT CPP_ArcShape {
         float EndAngle;
 
         uint64_t ID;
-        GLuint ColorIndex;
+        float ColorIndex;
         unsigned int Width = 0;
         unsigned int PointCount = 0;
         unsigned int Radius;
